refactor: create middleware specitif file for errorHandling, sessions, cors
